+++ juniper/src/tests/model.rs
@@ -17,7 +17,7 @@ pub trait Character {
     fn friend_ids(&self) -> &[String];
     fn appears_in(&self) -> &[Episode];
     fn secret_backstory(&self) -> &Option<String>;
-    fn as_character(&self) -> &Character;
+    fn as_character(&self) -> &dyn Character;
 }
 
 pub trait Human: Character {
@@ -62,7 +62,7 @@ impl Character for HumanData {
     fn secret_backstory(&self) -> &Option<String> {
         &self.secret_backstory
     }
-    fn as_character(&self) -> &Character {
+    fn as_character(&self) -> &dyn Character {
         self
     }
 }
@@ -89,7 +89,7 @@ impl Character for DroidData {
     fn secret_backstory(&self) -> &Option<String> {
         &self.secret_backstory
     }
-    fn as_character(&self) -> &Character {
+    fn as_character(&self) -> &dyn Character {
         self
     }
 }
@@ -248,7 +248,7 @@ impl Database {
         }
     }
 
-    pub fn get_hero(&self, episode: Option<Episode>) -> &Character {
+    pub fn get_hero(&self, episode: Option<Episode>) -> &dyn Character {
         if episode == Some(Episode::Empire) {
             self.get_human("1000").unwrap().as_character()
         } else {
@@ -256,15 +256,15 @@ impl Database {
         }
     }
 
-    pub fn get_human(&self, id: &str) -> Option<&Human> {
-        self.humans.get(id).map(|h| h as &Human)
+    pub fn get_human(&self, id: &str) -> Option<&dyn Human> {
+        self.humans.get(id).map(|h| h as &dyn Human)
     }
 
-    pub fn get_droid(&self, id: &str) -> Option<&Droid> {
-        self.droids.get(id).map(|d| d as &Droid)
+    pub fn get_droid(&self, id: &str) -> Option<&dyn Droid> {
+        self.droids.get(id).map(|d| d as &dyn Droid)
     }
 
-    pub fn get_character(&self, id: &str) -> Option<&Character> {
+    pub fn get_character(&self, id: &str) -> Option<&dyn Character> {
         if let Some(h) = self.humans.get(id) {
             Some(h)
         } else if let Some(d) = self.droids.get(id) {
@@ -274,7 +274,7 @@ impl Database {
         }
     }
 
-    pub fn get_friends(&self, c: &Character) -> Vec<&Character> {
+    pub fn get_friends(&self, c: &dyn Character) -> Vec<&dyn Character> {
         c.friend_ids()
             .iter()
             .flat_map(|id| self.get_character(id))